Move `Nimq_API` to `API\Helpers\Wallet`
